refactor Exception
