+++ src/Dacapo/Domain/Schema/IndexModifier/IndexModifierType/IndexModifierTypeFactory.php
@@ -2,7 +2,7 @@
 
 namespace UcanLab\LaravelDacapo\Dacapo\Domain\Schema\IndexModifier\IndexModifierType;
 
-use UcanLab\LaravelDacapo\Dacapo\Application\Shared\Exception\UseCase\InvalidArgumentException;
+use UcanLab\LaravelDacapo\Dacapo\Domain\Shared\Exception\Schema\IndexModifier\IndexModifierType\InvalidArgumentException;
 
 final class IndexModifierTypeFactory
 {
