+++ src/Dacapo/Domain/Shared/Exception/Schema/Column/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace UcanLab\LaravelDacapo\Dacapo\Domain\Shared\Exception\Schema\Column;
+
+use InvalidArgumentException as BaseException;
+
+final class InvalidArgumentException extends BaseException
+{
+}
