+++ src/Dacapo/Domain/Shared/Exception/Schema/ForeignKey/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace UcanLab\LaravelDacapo\Dacapo\Domain\Shared\Exception\Schema\ForeignKey;
+
+use InvalidArgumentException as BaseException;
+
+final class InvalidArgumentException extends BaseException
+{
+}
