+++ src/Dacapo/Domain/Shared/Exception/Schema/IndexModifier/IndexModifierType/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace UcanLab\LaravelDacapo\Dacapo\Domain\Shared\Exception\Schema\IndexModifier\IndexModifierType;
+
+use InvalidArgumentException as BaseException;
+
+final class InvalidArgumentException extends BaseException
+{
+}
